added frog enemies with pathfinding

+++ GameObject.cpp
@@ -24,8 +24,7 @@ GameObject::GameObject(Ogre::String newName, Ogre::SceneManager* scnMgr, Simulat
 }
 
 GameObject::~GameObject() {
-	printf("Calling destructor\n");
-    sceneMgr->destroySceneNode(rootNode);
+    if (rootNode != NULL) {sceneMgr->destroySceneNode(rootNode);}
     sceneMgr->destroyEntity(this->getName());
     simulator->getDynamicsWorld()->removeRigidBody(this->body);
     geom = NULL;
@@ -34,7 +33,7 @@ GameObject::~GameObject() {
     body = NULL;
     context = NULL;
     cCallBack = NULL;
-    simulator->printList();
+    //simulator->printList();
 }
 
 btRigidBody* GameObject::getBody() {
